+++ cmds/help.py
@@ -1,9 +1,8 @@
 from typing import Any
 
 import discord
+from core import Bot, CogInit, HelpMessages
 from discord.ext import commands
-
-from ..core import Bot, CogInit, HelpMessages
 
 
 class Help(CogInit):
@@ -21,7 +20,7 @@ class Help(CogInit):
         # Footer
         embed.set_footer(text=embed_info["footer"])
         # Thumbnail
-        embed.set_thumbnail(url=(await self.bot.fetch_guild(Bot.main_guild)).icon_url)
+        embed.set_thumbnail(url=(await self.bot.fetch_guild(Bot.active_guild)).icon_url)
         # Fields
         for command, description in embed_info["fields"]:
             embed.add_field(name=command, value=description, inline=True)
