+++ core/cog.py
